requirements_gpu.txt
numpy==1.16.0
pandas==0.24.0
tensorflow-gpu==1.12.0
matplotlib==3.0.2
scikit-learn==0.20.2
jupyter==1.0.0
Pillow==5.4.1
nltk==3.4
PyYAML==5.1